backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
pyjwt
psycopg
google-generativeai
mcp
